Sheepdog: Fix default sheepdog_store_address of glance-api.conf
The string of sheepdog_store_address is used by dog command
to specify the IP address of the sheep server.

In addition, dog command can't handle hostname and can handle
the strings of IP address only.

Previously the entry of glance-api.conf was 'localhost' and
dog command couldn't handle it.

So, this patch changes the entry from 'localhost' to the IP
address of localhost.

DocImpact

Change-Id: I601233fb75e2737a02e7af25a59062d26b31d994
Depends-On: I28ea18badcb4a47daece5efed38d53fc35870902